Big update
Added new pathing options for entity's
First up: pathfind.lua, probly quicker to look at the commenting than me explain that,

changed amost npc's that use pathing to use either basicPath or advancedPath, the advancedPath is the one that stands out as this allows the entity to pass id text and npc id's to automatically look at or speak at a a point in the path, and each point can have a delay.

Added random pathing for the bunnies in Khazam, as they just sat there.

I kept everything the same where i could, but i did have to redo some paths, but that let me cut down on path points also.