Adding capability for cross-server commands (lua)
- MSG_SEND_LUA_COMMAND - send a command to another server to run where the target is local
- MSG_REMOTE_PRINT_TO_PLAYER - send a message to a remote player for use with remote commands
- update message_server to lookup player server and redirect command/print
- update commandhandler::call to take a caller parameter
- add support for 't' command that will check if target is server local and if not redirect to message server
- add support in message for MSG_SEND_LUA_COMMAND/MSG_REMOTE_PRINT_TO_PLAYER
- message will now report if player is not found rather than scripts
- update packet_system to pass caller attribute for SmallPacket0x0B5 (chat)
- update luautils for RemotePrintToPlayer lua method
- hp.lua - example calling schema

It would require updates to all onTrigger in the /scripts/commands directory to take a first argument of caller. It would also be suggested to copy or make globally available the updated error and print_message that will attempt to handle whether or not player:PrintToPlayer or RemotePrintToPlayer should be used based on caller and player object ids matching (or not).

This is just a POC.