PR: Testing suite
The tests here are supposed to be at the integration level,
mostly just testing the interface into the new spell
stuff (scripts/global/spell.lua and all its supporting libs)(rename appropriately).

The tests are meant to be pretty non-commital, to make sure the interface
is working in general terms but not necessarily producing "correct"
numbers based on known constraints. This keeps the tests flexible
and useful for servers who may run slightly differe math (it is a pain
to ask them to refactor the test everytime they tweak their personal stuff).

The test setup includes support libraries for calls to the core, allowing
us to fake those without having to actually instantiate the core in our
testing of the lua side of things. Testing of functionality between the core
and lua should be done at a higher testing level than this.